My last implementation for the night. A commit for commenting may follow.
Functionality as it stands:
We can generate a factorbase of whatever size we say. Part of Algo Class.
An RArray object, which creates the testable Rs and delivers them to
    an RVal[] in Algo.
SolutionsProcessor object, which retrieves the solutions from Gaussian elimination
and searches for a factor. This implentation needs further testing after
    Gauss is implemented.
I made a Lamo pass at check the factors yielded by SolutionsProcessor, but
    haven't tested in the slightest.

Still need to...
    1) Implement the Gaussian elimination
    2) Find a way to make all this stuff parallel - it'll help efficiency
    3) Really check the factors, and possibly deliver them back to Main
    4) Vary j and k better - don't know how, but there has to be a better way